gamemanager/requirements.txt
PyJWT
flask
requests
